mailer/requirements.txt
Pillow
Flask
python-dotenv